ML/MLP/pyproject.toml
[tool.poetry]
name = "mlp-test"
version = "0.1.0"
description = ""
authors = ["Minseok Chu <mchu7797@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.0.1"
pillow = "^10.4.0"
matplotlib = "^3.9.1.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
